requirements.txt
cryptography>=41.0.0
msgpack>=1.0.5
argon2-cffi>=21.3.0
rich>=13.0.0
typer>=0.9.0
zstandard>=0.21.0
portalocker>=2.7.0
